Quartz: search tweets when interest is created

+++ src/main/java/com/fmalessio/twitterapi/service/InterestsServiceImpl.java
@@ -1,5 +1,7 @@
 package com.fmalessio.twitterapi.service;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -9,17 +11,28 @@ import com.fmalessio.twitterapi.repository.InterestRepository;
 @Service
 public class InterestsServiceImpl implements InterestsService {
 
-	@Autowired
 	private InterestRepository interestRepository;
+	private TweetsService tweetsService;
+
+	@Autowired
+	public InterestsServiceImpl(InterestRepository interestRepository, TweetsService tweetsService) {
+		this.interestRepository = interestRepository;
+		this.tweetsService = tweetsService;
+	}
 
 	@Override
 	public void create(Interest interest) {
-		interestRepository.saveAndFlush(interest);
+		interest = interestRepository.saveAndFlush(interest);
+		tweetsService.runScheduler(interest);
 	}
 
 	@Override
 	public void delete(long id) {
-		interestRepository.deleteById(id);
+		Optional<Interest> interest = interestRepository.findById(id);
+		if (interest.isPresent()) {
+			interestRepository.deleteById(id);
+			tweetsService.removeJobByInterestId(Long.toString(id));
+		}
 	}
 
 }
